Making OptimisticEmitFailureHandler class package-private and adding more tests.
